requirements.txt
Django~=4.2.9
djangorestframework~=3.14.0
Jinja2~=3.1.2
MarkupSafe~=2.1.3
asgiref~=3.7.2
certifi~=2023.11.17
charset-normalizer~=3.3.2
coreapi~=2.3.3
coreschema~=0.0.4
django-rest-swagger~=2.2.0
idna~=3.6
itypes~=1.2.0
openapi-codec~=1.3.2
pip~=23.3.2
pytz~=2023.3.post1
requests~=2.31.0
setuptools~=68.2.0
simplejson~=3.19.2
sqlparse~=0.4.4
typing-extensions~=4.9.0
tzdata~=2023.4
uritemplate~=4.1.1
urllib3~=2.1.0
wheel~=0.41.2